feat: componentize ban toggle button
